requirements.txt
# Base ----------------------------------------
sqlalchemy==1.4.37
pymysql==1.0.2
pandas==2.2.2
python-dotenv==0.20.0
psycopg2-binary==2.9.3
requests==2.27.1
spotipy==2.23.0

numpy>=1.18.5
opencv-python>=4.10.0.84
matplotlib>=3.9.2
ipyleaflet>=0.19.2
sympy>=1.13.2

# Machine learning ----------------------------
scikit-learn